separated spec files for feat grid
