+++ cypress/integration/v2/blocks/feature-grid/desktop-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopAdvanced } from '.'
+
+describe( 'Feature Grid Block ( Desktop Advanced )', registerTests( [ desktopAdvanced ] ) )
