+++ cypress/integration/v2/blocks/feature-grid/mobile-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileAdvanced } from '.'
+
+describe( 'Feature Grid Block ( Mobile Advanced )', registerTests( [ mobileAdvanced ] ) )
