+++ cypress/integration/v2/blocks/feature-grid/other-tests.spec.js
@@ -1,0 +1,23 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import {
+	blockExist,
+	blockError,
+	switchLayout,
+	switchDesign,
+	typeContent,
+} from '.'
+
+describe( 'Feature Grid Block ( Other Tests )', registerTests( [
+	blockExist,
+	blockError,
+	switchLayout,
+	switchDesign,
+	typeContent,
+] ) )
